(API) Move API classes to separate module
