+++ tests/test_rest.py
@@ -1,10 +1,8 @@
 import unittest.mock as mock
-import logging
 
 import pytest
-import testfixtures
 
-from sensormesh.rest import RestSource, RestTarget, RestApi, ApiMixin
+from sensormesh.rest import RestSource, RestTarget
 
 
 class TestRestSource:
@@ -101,42 +99,3 @@ class TestRestTarget:
         with o:
             o.update({'value': 1, 'other': 0})
         mock_api.post_update.assert_called_with({'field1': 1})
-
-
-class TestApiMixin:
-    def test_can_use_object_as_context_manager(self):
-        mock_api = RestApi()
-        obj = ApiMixin(api=mock_api, name='mock_obj')
-
-        with testfixtures.LogCapture(level=logging.DEBUG) as l:
-            with obj as context:
-                # Check that context is the same object
-                assert context is obj
-
-                assert len(l.records) == 2
-                assert_record_is(l.records[0], 'INFO', "Opening ApiMixin(name='mock_obj')")
-                assert_record_is(l.records[1], 'INFO', "Opening RestApi()")
-
-        assert len(l.records) == 4
-        assert_record_is(l.records[2], 'INFO', "Closing RestApi()")
-        assert_record_is(l.records[3], 'INFO', "Closing ApiMixin(name='mock_obj')")
-
-    def test_can_open_and_close_object(self):
-        mock_api = RestApi()
-        obj = ApiMixin(api=mock_api, name='mock_obj')
-
-        with testfixtures.LogCapture(level=logging.DEBUG) as l:
-            obj.open()
-            assert len(l.records) == 2
-            assert_record_is(l.records[0], 'INFO', "Opening ApiMixin(name='mock_obj')")
-            assert_record_is(l.records[1], 'INFO', "Opening RestApi()")
-
-            obj.close()
-            assert len(l.records) == 4
-            assert_record_is(l.records[2], 'INFO', "Closing RestApi()")
-            assert_record_is(l.records[3], 'INFO', "Closing ApiMixin(name='mock_obj')")
-
-
-def assert_record_is(record, level, message):
-    assert record.levelname == level
-    assert record.getMessage() == message
